+++ ePadel/ePadel.Model/Tereni.cs
@@ -15,7 +15,7 @@ namespace ePadel.Model
 
         public int BrojTerena { get; set; }
 
-        public decimal Cijena { get; set; }
+        public int Cijena { get; set; }
 
         public int? TipTerenaId { get; set; }
         public virtual TipTerena? TipTerena { get; set; }
