+++ ePadel/ePadel.Services/Database/Tereni.cs
@@ -8,7 +8,7 @@ namespace ePadel.Services.Database
         public int TerenId { get; set; }
         public string? Naziv { get; set; }
         public int? BrojTerena { get; set; }
-        public decimal? Cijena { get; set; }
+        public int? Cijena { get; set; }
         public int? TipTerenaId { get; set; }
 
         public virtual TipTerena? TipTerena { get; set; }
